% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalmix_funcs.R
\name{anFormula}
\alias{anFormula}
\title{anFormula}
\usage{
anFormula(par, m, n)
}
\arguments{
\item{par}{The parameter estimates as a list containing alpha, mu, sigma, and gamma 
in the form of (alpha = (alpha_1, ..., alpha_m), mu = (mu_1, ..., mu_m), 
sigma = (sigma_1, ..., sigma_m), gamma = (gamma_1, ..., gamma_m))
Computes a_n based on empirical results found in Kasahara and Shimotsu (2015)}

\item{m}{The number of components in the mixture}

\item{n}{The number of observations}

\item{par}{The parameter estimates as a list containing alpha, mu, sigma, and gamma 
in the form of (alpha = (alpha_1, ..., alpha_m), mu = (mu_1, ..., mu_m), 
sigma = (sigma_1, ..., sigma_m), gamma = (gamma_1, ..., gamma_m))}
}
\value{
a_n used to initialize values
}
\references{
Kasahara, H., and Shimotsu, K. (2015)
Testing the Number of Components in Normal Mixture Regression Models,
\emph{Journal of the American Statistical Association},
\bold{110}, 1632--1645.
}

