% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getComponentcomponents}
\alias{getComponentcomponents}
\title{getComponentcomponents}
\usage{
getComponentcomponents(postprobs)
}
\arguments{
\item{postprobs}{n by m matrix of posterior probabilities for
m-component model on n observations}
}
\value{
n by 1 vector of components that indicate which component each observation belongs to
based on its posterior probability
}
\description{
Generates a vector that indicates which component each observation belongs to,
based on its posterior probability
}

