% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalmix_funcs.R
\name{normalmixCritBoot}
\alias{normalmixCritBoot}
\title{normalmixCritBoot}
\usage{
normalmixCritBoot(y, parlist, z = NULL, values = NULL, ninits = 10,
  nbtsp = 199, parallel = TRUE, cl = NULL)
}
\arguments{
\item{y}{n by 1 vector of data}

\item{z}{n by p matrix of regressor associated with gamma}

\item{values}{vector of length 3 (k = 1, 2, 3) at which the p-values are computed}

\item{ninits}{}

\item{nbtsp}{The number of bootstrap observations. By default, it is set to be 199.}

\item{parallel}{Determines whether package \code{parallel} is used to compute p-values. 
By default, it is set \code{TRUE}.}

\item{cl}{Cluster used for parallelization (optional)}

\item{coefficients}{(alpha_1, ..., alpha_m, mu_1, ..., mu_m, sigma_1, ..., sigma_m, gamma)}
}
\value{
A list with the following items:
\item{crit}{3 by 3 matrix of (0.1, 0.05, 0.01 critical values), jth row corresponding to k=j}
\item{pvals}{A vector of p-values at k = 1, 2, 3}
}
\description{
Computes the bootstrap critical values of the modified EM test.
}

