% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalmix_funcs.R
\name{normalmixMaxPhiStep}
\alias{normalmixMaxPhiStep}
\title{normalmixMaxPhiStep}
\usage{
normalmixMaxPhiStep(htaupair, y, par0, z = NULL, p, jpvt, an, ninits,
  ninits.short, epsilon.short, epsilon, maxit.short, maxit, verb)
}
\arguments{
\item{htaupair}{A set of h and tau}

\item{y}{n by 1 vector of data}

\item{par0}{A list of parameters for initialization (of class \code{normalMix})}

\item{z}{n by p matrix of regressor associated with gamma}

\item{p}{Dimension of z}

\item{jpvt}{Pivots used in dgelsy}

\item{an}{a term used for penalty function}

\item{ninits}{The number of randomly drawn initial values.}

\item{ninits.short}{The number of candidates used to generate an initial phi, in short MEM}

\item{epsilon.short}{The convergence criterion in short EM. Convergence is declared when the penalized log-likelihood increases by less than \code{epsilon.short}.}

\item{epsilon}{The convergence criterion. Convergence is declared when the penalized log-likelihood increases by less than \code{epsilon}.}

\item{maxit.short}{The maximum number of iterations in short EM.}

\item{maxit}{The maximum number of iterations.}
}
\value{
A list of phi, log-likelihood, and penalized log-likelihood resulting from MEM algorithm.
}
\description{
Given a pair of h and tau and data, compute ordinary & 
Penalized log-likelihood ratio resulting from MEM algorithm at k=1,2,3, tailored for parallelization.
}

