% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalmix_funcs.R
\name{normalmixMEMtestSeq}
\alias{normalmixMEMtestSeq}
\title{normalmixMEMtestSeq}
\usage{
normalmixMEMtestSeq(y, z = NULL, maxm = 3, ninits = 10, maxit = 2000,
  nbtsp = 199, parallel = FALSE, cl = NULL)
}
\arguments{
\item{y}{n by 1 vector of data for y}

\item{z}{n by p matrix of regressor associated with gamma}

\item{maxm}{The maximum number of components set as null hypothesis in the mixture}

\item{ninits}{The number of randomly drawn initial values.}

\item{maxit}{The maximum number of iterations.}

\item{nbtsp}{The number of bootstrap observations; by default, it is set to be 199}

\item{parallel}{Determines whether package \code{doParallel} is used for calculation}

\item{cl}{Cluster used for parallelization; if it is \code{NULL}, the system will automatically
create a new one for computation accordingly.}
}
\value{
A list of with the following items:
\item{alpha}{maxm by maxm matrix, whose i-th column is a vector of alphas estimated given the null hypothesis m_0 = i}
\item{mu}{maxm by maxm matrix, whose i-th column is a vector of mus estimated given the null hypothesis m_0 = i}
\item{sigma}{maxm by maxm matrix, whose i-th column is a vector of sigmas estimated given the null hypothesis m_0 = i}
\item{beta}{A list of length maxm, whose i-th element is a q times i matrix of betas estimated given the null hypothesis m_0 = i}
\item{gamma}{maxm by maxm matrix, whose i-th column is a vector of gammas estimated given the null hypothesis m_0 = i}
\item{emstat}{A maxm vector of values of modified EM statistics of the model at m_0 = 1, 2, ..., maxm}
\item{pvals}{A maxm by 3 matrix whose i-th row indicates a vector of p-values at k = 1, 2, 3}
\item{aic}{A maxm vector of Akaike Information Criterion of the fitted model at m_0 = 1, 2, ..., maxm}
\item{bic}{A maxm vector of Bayesian Information Criterion of the fitted model at m_0 = 1, 2, ..., maxm}
\item{loglik}{A maxm vector of log-likelihood values of the model at m_0 = 1, 2, ..., maxm}
\item{penloglik}{A maxm vector of penalized log-likelihood values of the model at m_0 = 1, 2, ..., maxm}
}
\description{
Sequentially performs MEM test given the data for y and x on the null hypothesis H_0: m = m_0
where m_0 is in {1, 2, ..., maxm}
}
\examples{
data(faithful)
attach(faithful)
normalmixMEMtestSeq(y = eruptions)
}

