% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalmix_funcs.R
\name{normalmixPhi2}
\alias{normalmixPhi2}
\title{normalmixPhi2}
\usage{
normalmixPhi2(y, z = NULL, par0, sigma0, h, tau, an)
}
\arguments{
\item{y}{n by 1 vector of data}

\item{z}{n by p matrix of regressor associated with gamma}

\item{par0}{A list of parameters for initialization (of class \code{normalMix})}

\item{sigma0}{m by 1 vector of sigma}

\item{h}{h used as index for pivoting}

\item{tau}{Tau used to split the h-th component}

\item{an}{a term used for penalty function}

\item{htaupair}{A set of h and tau}
}
\value{
A list of length 2, where the first element is the local modified 
MEM test statistic at k=2, the second element is 
the local modified MEM test statistic at k=3
}
\description{
Starting from a parameter estimate, take two EM steps to compute the
local modified EM test statistic for testing H_0 of m components
against H_1 of m+1 at K=2,3 for a univariate normal finite mixture
}

