% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalmix_funcs.R
\name{normalmixPhiInit}
\alias{normalmixPhiInit}
\title{normalmixPhiInit}
\usage{
normalmixPhiInit(y, parlist, z = NULL, h, tau, ninits = 1)
}
\arguments{
\item{y}{n by 1 vector of data}

\item{parlist}{The parameter estimates as a list containing alpha, mu, sigma, and gamma 
in the form of (alpha = (alpha_1, ..., alpha_m), mu = (mu_1, ..., mu_m), 
sigma = (sigma_1, ..., sigma_m), gamma = (gamma_1, ..., gamma_m))}

\item{z}{n by p matrix of regressor associated with gamma}

\item{h}{h used as index for pivoting}

\item{tau}{Tau used to split the h-th component}

\item{ninits}{number of initial values to be generated}

\item{z}{n by p matrix of regressor associated with gamma}
}
\value{
A list with the following items:
\item{alpha}{m+1 by ninits matrix for alpha}
\item{mu}{m+1 by ninits matrix for mu}
\item{sigma}{m+1 by ninits matrix for sigma}
\item{gamma}{m+1 by ninits matrix for gamma}
}
\description{
Generates initial values used by the modified EM test for mixture of normals.
}

