% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regmix_funcs.R
\name{obj_zIz.jac}
\alias{obj_zIz.jac}
\title{Computes Jacobian of the objective function used in computing LR_2}
\usage{
obj_zIz.jac(b, Z, I)
}
\description{
Computes Jacobian of the objective function used in computing LR_2
}

