% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalmix_funcs.R
\name{omega.12}
\alias{omega.12}
\title{omega.12}
\usage{
omega.12(parlist)
}
\arguments{
\item{parlist}{The parameter estimates as a list containing alpha, mu, sigma, and gamma 
in the form of (alpha = (alpha_1, ..., alpha_m), mu = (mu_1, ..., mu_m), 
sigma = (sigma_1, ..., sigma_m), gamma = (gamma_1, ..., gamma_m))}
}
\value{
The misclassification rate omega_ij
}
\description{
Computes omega_{12} for testing H_0: m=2 against H_1: m=3 defined in Maitra and Melnykov (2010)
}
\references{
Maitra, R., and Melnykov, V. (2010)
Simulating Data to Study Performance of Finite Mixture Modeling and Model-Based Clustering Algorithms,
\emph{Journal of Computational and Graphical Statistica},
\bold{19}, 354--376.
}

