% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalmix_funcs.R
\name{omega.123}
\alias{omega.123}
\title{omega.123}
\usage{
omega.123(parlist)
}
\arguments{
\item{parlist}{The parameter estimates as a list containing alpha, mu, sigma, and gamma
in the form of (alpha = (alpha_1, ..., alpha_m), mu = (mu_1, ..., mu_m),
sigma = (sigma_1, ..., sigma_m), gam = (gamma_1, ..., gamma_m))}
}
\value{
A 2 by 1 vector whose first element is omega_12 and second element is omega_23
}
\description{
Computes omega_{12} and omega_{23} defined in Maitra and Melnykov (2010)
}
\references{
Maitra, R., and Melnykov, V. (2010)
Simulating Data to Study Performance of Finite Mixture Modeling and Model-Based Clustering Algorithms,
\emph{Journal of Computational and Graphical Statistica},
\bold{19}, 354--376.
}

