% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalmix_funcs.R
\name{omega.ji}
\alias{omega.ji}
\title{omega.ji}
\usage{
omega.ji(phi_i, phi_j)
}
\arguments{
\item{phi_i}{3 by 1 column consisting of alpha, mu, sigma of ith component}

\item{phi_j}{3 by 1 column consisting of alpha, mu, sigma of jth component}
}
\value{
omega_{j|i}
}
\description{
Computes omega_{j|i} defined in (2.1) of Maitra and Melnykov (2010)
}
\references{
Maitra, R., and Melnykov, V. (2010)
Simulating Data to Study Performance of Finite Mixture Modeling and Model-Based Clustering Algorithms,
\emph{Journal of Computational and Graphical Statistica},
\bold{19}, 354--376.
}

