% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plotDiag}
\alias{plotDiag}
\title{plotDiag}
\usage{
plotDiag(components, y = y, x = x, m = 2)
}
\arguments{
\item{components}{n vector of the indices of components for each observation}

\item{y}{n vector of data that represents dependent variables}

\item{x}{n by q matrix of data that represent(s) covariates}

\item{m}{The number of components}
}
\description{
Prints the multiple plots for diagnosis.
}
\examples{
data(faithful)
attach(faithful)
regmix.model.m2 <- regmixPMLE(y = eruptions, x = waiting, m = 2)
\dontrun{plotDiag(regmix.model.m2$components, y = eruptions, x = waiting, m  = 2)}
}

