% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.normalregMix}
\alias{print.normalregMix}
\title{print.normalregMix}
\usage{
\method{print}{normalregMix}(x, ...)
}
\arguments{
\item{x}{normalregMix instance}

\item{...}{Other arguments that do not affect the function}
}
\description{
Prints the description of a normalregMix instance
}

