% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.normalregMix}
\alias{print.summary.normalregMix}
\title{print.summary.normalregMix}
\usage{
\method{print}{summary.normalregMix}(x, ...)
}
\arguments{
\item{x}{normalregMix instance}

\item{...}{Other arguments that do not affect the function}
}
\description{
Prints the summary of a normalregMix instance
}

