% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regmix_funcs.R
\name{regmixCritBoot}
\alias{regmixCritBoot}
\title{regmixCritBoot}
\usage{
regmixCritBoot(y, x, parlist, z = NULL, values = NULL, ninits = 100,
  nbtsp = 199, parallel = TRUE, cl = NULL)
}
\arguments{
\item{y}{n by 1 vector of data for y}

\item{x}{n by q vector of data for x}

\item{parlist}{The parameter estimates as a list containing alpha, mu, sigma, and gamma 
in the form of (alpha = (alpha_1, ..., alpha_m), mu = (mu_1, ..., mu_m), 
sigma = (sigma_1, ..., sigma_m), gamma = (gamma_1, ..., gamma_m))}

\item{z}{n by p matrix of regressor associated with gamma}

\item{values}{3 by 1 Vector of length 3 (k = 1, 2, 3) at which the p-values are computed}

\item{ninits}{The number of initial candidates to be generated}

\item{nbtsp}{The number of bootstrap observations; by default, it is set to be 199}

\item{parallel}{Determines whether package \code{parallel} is used to compute p-values. 
By default, it is set \code{TRUE}.}

\item{cl}{Cluster used for parallelization (optional)}
}
\value{
A list with the following items:
\item{crit}{3 by 3 matrix of (0.1, 0.05, 0.01 critical values), jth row corresponding to k=j}
\item{pvals}{A vector of p-values at k = 1, 2, 3}
}
\description{
Computes the bootstrap critical values of the modified EM test.
}

