% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regmix_tests.R
\name{regmixMEMtest}
\alias{regmixMEMtest}
\title{regmixMEMtest}
\usage{
regmixMEMtest(y, x, m = 2, z = NULL, tauset = c(0.1, 0.3, 0.5),
  an = NULL, ninits = 100, crit.method = c("none", "asy", "boot"),
  nbtsp = 199, cl = NULL, parallel = FALSE)
}
\arguments{
\item{y}{n by 1 vector of data for y}

\item{x}{n by q matrix of data for x}

\item{m}{The number of components in the mixture defined by a null hypothesis, m_0}

\item{z}{n by p matrix of regressor associated with gamma}

\item{tauset}{A set of initial tau value candidates}

\item{an}{a term used for penalty function}

\item{ninits}{The number of randomly drawn initial values.}

\item{crit.method}{Method used to compute the variance-covariance matrix,
one of \code{"none"}, \code{"asy"}, and \code{"boot"}. The default option is
'\code{"none"}. When \code{method = "asy"}, the p-values are computed
based on an asymptotic method. When \code{method = "OPG"},
the p-values are generated by bootstrapping.}

\item{nbtsp}{The number of bootstrap observations; by default, set as 199.}

\item{cl}{Cluster used for parallelization; if it is \code{NULL}, the system
will automatically generate a new one for computation accordingly.}

\item{parallel}{Determines whether package \code{doParallel} is used for calculation}
}
\value{
A list of class \code{normalMix} with items:
\item{coefficients}{A vector of parameter estimates. Ordered as \eqn{
'\alpha_1,\ldots,\alpha_m,\mu_1,\ldots,\mu_m,\sigma_1,\ldots,\sigma_m,\gam}.}
\item{parlist}{The parameter estimates as a list containing alpha, mu, and
sigma (and gamma if z is included in the model).}
\item{vcov}{The estimated variance-covariance matrix.}
\item{loglik}{The maximized value of the log-likelihood.}
\item{penloglik}{The maximized value of the penalized log-likelihood.}
\item{aic}{Akaike Information Criterion of the fitted model.}
\item{bic}{Bayesian Information Criterion of the fitted model.}
\item{postprobs}{n by m matrix of posterior probabilities for observations}
\item{components}{n by 1 vector of integers that indicates the indices of components
each observation belongs to based on computed posterior probabilities}
\item{call}{The matched call.}
\item{m}{The number of components in the mixture.}
}
\description{
Performs MEM test given the data for y and x on 
the null hypothesis H_0: m = m_0. Using this function is equivalent to 
calling normalmixMEMtest with regressors specified by x as a parameter.
}
\examples{
data(faithful)
attach(faithful)
\dontrun{regmixMEMtest(y = eruptions, x = waiting, m = 1, crit.method = "asy")}
\dontrun{regmixMEMtest(y = eruptions, x = waiting, m = 2, parallel = TRUE, crit.method = "asy")}
}

