% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regmix_funcs.R
\name{regmixPMLEinit}
\alias{regmixPMLEinit}
\title{regmixPMLEinit}
\usage{
regmixPMLEinit(y, x, z = NULL, ninits = 1, m = 2)
}
\arguments{
\item{y}{n by 1 vector of data for y}

\item{x}{n by q matrix of data for x}

\item{z}{n by p matrix of regressor associated with gam}

\item{ninits}{number of initial values to be generated}

\item{m}{The number of components in the mixture}
}
\value{
A list with the following items:
\item{alpha}{m by ninits matrix for alpha}
\item{mubeta}{q+1 by m by ninits array for mu and beta}
\item{sigma}{m by ninits matrix for sigma}
\item{gam}{m by ninits matrix for gam}
}
\description{
Generate initial values used by the PMLE of mixture of normals
}

