% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regmix_funcs.R
\name{regmixVcov}
\alias{regmixVcov}
\title{regmixVcov}
\usage{
regmixVcov(y, x, coefficients, z = NULL, vcov.method = c("Hessian", "OPG"))
}
\arguments{
\item{y}{n by 1 vector of data for y}

\item{x}{n by q matrix of data for x}

\item{coefficients}{(alpha_1, ..., alpha_m, mu_1, ..., mu_m, sigma_1, ..., sigma_m, gamma)}

\item{z}{n by p matrix of regressor associated with gamma}

\item{vcov.method}{Method used to compute the variance-covariance matrix, 
one of \code{"Hessian"} and \code{"OPG"}. #' The default option is \code{"Hessian"}. 
When \code{method = "Hessian"}, the variance-covarince matrix is 
estimated by the Hessian using the formula given in Boldea and Magnus (2009). 
When \code{method = "OPG"}, the outer product of gradients is used.}

\item{p}{Dimension of z}
}
\value{
The variance-covariance matrix of the MLE of 
m-component normal mixture given the data and coefficients.
}
\description{
Computes the variance-covariance matrix of the MLE of m-component normal mixture.
}
\references{
Boldea, O. and Magnus, J. R. (2009)
Maximum Likelihood Estimation of the Multivariate Normal Mixture Model,
\emph{Journal of the American Statistical Association},
\bold{104}, 1539--1549.
}

