% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.normalregMix}
\alias{summary.normalregMix}
\title{summary.normalregMix}
\usage{
\method{summary}{normalregMix}(object, reorder = FALSE, digits = 3, ...)
}
\arguments{
\item{object}{normalregMix instance}

\item{reorder}{Determines whether components are reordered in summary}

\item{digits}{Digits used for reporting}

\item{...}{Other arguments that do not affect the function}
}
\description{
Returns the summary of a normalregMix instance
}

