% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalmix_funcs.R
\name{normalmixCritBoot}
\alias{normalmixCritBoot}
\title{normalmixCritBoot}
\usage{
normalmixCritBoot(y, parlist, z = NULL, values = NULL, ninits = 10,
  nbtsp = 199, parallel = TRUE, cl = NULL)
}
\arguments{
\item{y}{n by 1 vector of data}

\item{parlist}{The parameter estimates as a list containing alpha, mu, sigma, and gamma 
in the form of (alpha = (alpha_1, ..., alpha_m), mu = (mu_1, ..., mu_m), 
sigma = (sigma_1, ..., sigma_m), gamma = (gamma_1, ..., gamma_m))}

\item{z}{n by p matrix of regressor associated with gamma}

\item{values}{3 by 1 Vector of length 3 (k = 1, 2, 3) at which the p-values are computed}

\item{ninits}{The number of initial candidates to be generated}

\item{nbtsp}{The number of bootstrap observations; by default, it is set to be 199.}

\item{parallel}{Determines whether package \code{doParallel} is used for calculation}

\item{cl}{Cluster used for parallelization (optional)}
}
\value{
A list with the following items:
\item{crit}{3 by 3 matrix of (0.1, 0.05, 0.01 critical values), jth row corresponding to k=j}
\item{pvals}{A vector of p-values at k = 1, 2, 3}
}
\description{
Computes the bootstrap critical values of the modified EM test.
}

