% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regmix_funcs.R
\name{regmixPhiStep}
\alias{regmixPhiStep}
\title{regmixPhiStep}
\usage{
regmixPhiStep(htaupair, y, x, parlist, z = NULL, p, an, ninits, ninits.short,
  epsilon.short, epsilon, maxit.short, maxit, verb)
}
\arguments{
\item{htaupair}{A set of h and tau}

\item{y}{n by 1 vector of data for y}

\item{x}{n by q matrix of data for x}

\item{parlist}{The parameter estimates as a list containing alpha, mu, sigma, and gam
in the form of (alpha = (alpha_1, ..., alpha_m), mu = (mu_1, ..., mu_m),
sigma = (sigma_1, ..., sigma_m), gam = (gam_1, ..., gam_m))}

\item{z}{n by p matrix of regressor associated with gam}

\item{p}{Dimension of z}

\item{an}{a term used for penalty function}

\item{ninits}{The number of randomly drawn initial values.}

\item{ninits.short}{The number of candidates used to generate an initial phi, in short MEM}

\item{epsilon.short}{The convergence criterion in short EM. Convergence is declared when the penalized log-likelihood increases by less than \code{epsilon.short}.}

\item{epsilon}{The convergence criterion. Convergence is declared when the penalized log-likelihood increases by less than \code{epsilon}.}

\item{maxit.short}{The maximum number of iterations in short EM.}

\item{maxit}{The maximum number of iterations.}

\item{verb}{Determines whether to print a message if an error occurs.}
}
\value{
A list of coefficients, log-likelihood, and penalized log-likelihood resulting from MEM algorithm.
}
\description{
Given a pair of h and tau and data, compute ordinary &
penalized log-likelihood ratio resulting from MEM algorithm at k=1,2,3, 
tailored for parallelization.
}

